+++ middleware/validation.js
@@ -37,6 +37,47 @@ const checkFieldLength = (route, fields) => {
   return false;
 };
 
+// checking for undefined fields
+const undefinedFields = (req, res) => {
+  const { username, bio } = req.body;
+
+  if (username === undefined || bio === undefined) {
+    return res.status(400).json({
+      status: 'fail',
+      message: 'All fields are required'
+    });
+  }
+};
+
+// checking for any empty field
+const emptyField = (req, res) => {
+  const { username, bio } = req.body;
+
+  if (!username.trim() || !bio.trim()) {
+    return res.status(400).json({
+      status: 'fail',
+      message: 'Fields cannot be empty'
+    });
+  }
+};
+
+// checking for any unwanted field
+const extraFields = (req, res) => {
+  const fieldLength = Object.keys(req.body).length;
+  if (fieldLength !== 2) {
+    return res.status(400).json({
+      status: 'fail',
+      message: 'Extra field(s) not required'
+    });
+  }
+};
+
+exports.profileValidation = (req, res) => {
+  undefinedFields(req, res);
+  emptyField(req, res);
+  extraFields(req, res);
+};
+
 exports.validate = route => (req, res, next) => {
   const userDetails = req.body;
   const tooManyFields = checkFieldLength(route, userDetails);
